requirements.txt
scikit-learn
pandas
numpy
matplotlib
tensorflow
logparser3
tqdm
sentence-transformers